+++ t/db_dependent/RapidoILL/QueuedTasks.t
@@ -1,21 +1,19 @@
 #!/usr/bin/perl
 
-# Copyright 2025 ByWater Solutions
+# This file is part of the Rapido ILL plugin
 #
-# This file is part of Koha.
-#
-# Koha is free software; you can redistribute it and/or modify it
+# The Rapido ILL plugin is free software; you can redistribute it and/or modify it
 # under the terms of the GNU General Public License as published by
 # the Free Software Foundation; either version 3 of the License, or
 # (at your option) any later version.
 #
-# Koha is distributed in the hope that it will be useful, but
+# The Rapido ILL plugin is distributed in the hope that it will be useful, but
 # WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with Koha; if not, see <http://www.gnu.org/licenses>.
+# along with The Rapido ILL plugin; if not, see <https://www.gnu.org/licenses>.
 
 use Modern::Perl;
 
@@ -300,7 +298,7 @@ subtest 'Complex filtering and method chaining' => sub {
     $schema->storage->txn_begin;
 
     my $initial_count = RapidoILL::QueuedTasks->search()->count;
-    my $tasks = RapidoILL::QueuedTasks->new;
+    my $tasks         = RapidoILL::QueuedTasks->new;
 
     # Create a variety of tasks
     for my $i ( 1 .. 10 ) {
@@ -387,8 +385,8 @@ subtest 'enqueue() context fallback tests' => sub {
     );
 
     my $default_context = $task_default->decoded_context;
-    ok( $default_context, 'Default context created' );
-    ok( exists $default_context->{userenv}, 'Default context has userenv' );
+    ok( $default_context,                     'Default context created' );
+    ok( exists $default_context->{userenv},   'Default context has userenv' );
     ok( exists $default_context->{interface}, 'Default context has interface' );
 
     $schema->storage->txn_rollback;